Done some debugging and wrote integration tests

+++ modules/rsync/post-check.7.py
@@ -1,7 +1,7 @@
 import os
 import datetime
 
-def get_day():
+def __get_day(index):
     trad = {
             0: 'Monday',
             1: 'Tuesday',
@@ -12,7 +12,10 @@ def get_day():
             6: 'Sunday'
     }
 
-    return trad.get(datetime.datetime.today().weekday())
+    return trad.get(index)
+
+def get_day():
+    return __get_day(datetime.datetime.now().weekday())
 
 def is_mounted():
     handle = os.popen('ssh root@bach.dest.1 df')
@@ -21,23 +24,32 @@ def is_mounted():
     pos = contents.find('/root/test2')
     return pos != -1
 
+def list_files():
+    handle = os.popen('ssh root@bach.dest.1 ls -l /root/test2')
+    contents = handle.read()
+    handle.close()
+    print(contents)
+
 if __name__ == '__main__':
+    follow = True
     if is_mounted():
         os._exit(1)
-    raw = os.system('ssh root@bach.dest.1 mount -o loop,offset=1048576 /root/test.img /root/test2')
+    raw = os.system('ssh root@bach.dest.1 mount -o loop,offset=1048576 /root/test2.img /root/test2')
     stat = os.waitstatus_to_exitcode(raw)
     if stat != 0:
-        os._exit(1)
+        print("Unable to mount /root/test2.img")
+        follow = False
+    list_files()
+    if follow:
+        raw = os.system('ssh root@bach.dest.1 diff -r /root/test2/{} /root/compare'.format(get_day()))
+        stat = os.waitstatus_to_exitcode(raw)
+        if stat < 0:
+            stat = stat * -1
 
-    raw = os.system('ssh root@bach.dest.1 diff -r /root/test2/{} /root/compare'.format(get_day()))
-    stat = os.waitstatus_to_exitcode(raw)
-    if stat < 0:
-        stat = stat * -1
-    
-    os.system('ssh root@bach.dest.1 rm -r /root/test2/*')
-    for i in range(6):
-        os.system('ssh root@bach.dest.1 mkdir /root/test2/{}'.format(get_day()))
+        os.system('ssh root@bach.dest.1 rm -r /root/test2/*')
+        for i in range(0,7):
+            os.system('ssh root@bach.dest.1 mkdir -p /root/test2/{}'.format(__get_day(i)))
 
-    os.system('ssh root@bach.dest.1 umount /root/test2')
+        os.system('ssh root@bach.dest.1 umount /root/test2')
 
     os._exit(stat)
